Add CategoryManager for method addAndPruneAllOptionCombos.
