+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/category/CategoryService.java
@@ -427,20 +427,6 @@ public interface CategoryService
     void updateOptionCombos( CategoryCombo categoryCombo );
 
     /**
-     * Generates the complete set of category option combos for the given
-     * category combo. Removes obsolete category option combos.
-     *
-     * @param categoryCombo the CategoryCombo.
-     */
-    void addAndPruneOptionCombos( CategoryCombo categoryCombo );
-
-    /**
-     * Generates the complete set of category option combos for all category
-     * combos.
-     */
-    void addAndPruneAllOptionCombos();
-
-    /**
      * Returns the category option combo with the given uid. Respects access control
      * by only returning objects which all category options are accessible.
      *
