+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/META-INF/dhis/beans.xml
@@ -904,6 +904,8 @@
 
   <bean id="validationNotificationMessageRenderer" class="org.hisp.dhis.notification.ValidationNotificationMessageRenderer" />
 
+  <bean id="org.hisp.dhis.category.CategoryManager" class="org.hisp.dhis.category.DefaultCategoryManager"/>
+
   <bean id="org.hisp.dhis.category.CategoryService" class="org.hisp.dhis.dataelement.DefaultCategoryService">
     <property name="categoryStore" ref="org.hisp.dhis.category.CategoryStore" />
     <property name="categoryOptionStore" ref="org.hisp.dhis.category.CategoryOptionStore" />
